+++ ios/RCTConvert+BitmovinPlayer.swift
@@ -636,78 +636,40 @@ extension RCTConvert {
      - Parameter json: JS object.
      - Returns: The associated `BitmovinAnalyticsConfig` value or nil.
      */
-    static func analyticsConfig(_ json: Any?) -> BitmovinAnalyticsConfig? {
+    static func analyticsConfig(_ json: Any?) -> AnalyticsConfig? {
         guard
             let json = json as? [String: Any?],
-            let key = json["key"] as? String
+            let key = json["licenseKey"] as? String
         else {
             return nil
         }
-        let config: BitmovinAnalyticsConfig
-        if let playerKey = json["playerKey"] as? String {
-            config = BitmovinAnalyticsConfig(key: key, playerKey: playerKey)
-        } else {
-            config = BitmovinAnalyticsConfig(key: key)
-        }
-        if let cdnProvider = json["cdnProvider"] as? String {
-            config.cdnProvider = cdnProvider
-        }
-        if let customerUserId = json["customUserId"] as? String {
-            config.customerUserId = customerUserId
-        }
-        if let experimentName = json["experimentName"] as? String {
-            config.experimentName = experimentName
-        }
-        if let videoId = json["videoId"] as? String {
-            config.videoId = videoId
-        }
-        if let title = json["title"] as? String {
-            config.title = title
-        }
-        if let path = json["path"] as? String {
-            config.path = path
-        }
-        if let isLive = json["isLive"] as? Bool {
-            config.isLive = isLive
-        }
-        if let ads = json["ads"] as? Bool {
-            config.ads = ads
-        }
-        if let randomizeUserId = json["randomizeUserId"] as? Bool {
-            config.randomizeUserId = randomizeUserId
-        }
-        config.customData1 = json["customData1"] as? String
-        config.customData2 = json["customData2"] as? String
-        config.customData3 = json["customData3"] as? String
-        config.customData4 = json["customData4"] as? String
-        config.customData5 = json["customData5"] as? String
-        config.customData6 = json["customData6"] as? String
-        config.customData7 = json["customData7"] as? String
-        config.customData8 = json["customData8"] as? String
-        config.customData9 = json["customData9"] as? String
-        config.customData10 = json["customData10"] as? String
-        config.customData11 = json["customData11"] as? String
-        config.customData12 = json["customData12"] as? String
-        config.customData13 = json["customData13"] as? String
-        config.customData14 = json["customData14"] as? String
-        config.customData15 = json["customData15"] as? String
-        config.customData16 = json["customData16"] as? String
-        config.customData17 = json["customData17"] as? String
-        config.customData18 = json["customData18"] as? String
-        config.customData19 = json["customData19"] as? String
-        config.customData20 = json["customData20"] as? String
-        config.customData21 = json["customData21"] as? String
-        config.customData22 = json["customData22"] as? String
-        config.customData23 = json["customData23"] as? String
-        config.customData24 = json["customData24"] as? String
-        config.customData25 = json["customData25"] as? String
-        config.customData26 = json["customData26"] as? String
-        config.customData27 = json["customData27"] as? String
-        config.customData28 = json["customData28"] as? String
-        config.customData29 = json["customData29"] as? String
-        config.customData30 = json["customData30"] as? String
-        config.experimentName = json["experimentName"] as? String
+        let randomizeUserId = json["randomizeUserId"] as? Bool
+        let adTrackingDisabled = json["adTrackingDisabled"] as? Bool
+   
+        let config = AnalyticsConfig(
+            licenseKey: key,
+            randomizeUserId: randomizeUserId ?? false,
+            adTrackingDisabled: adTrackingDisabled ?? false
+        )
         return config
+    }
+    
+    static func analyticsDefaultMetadataFromAnalyticsConfig(_ json: Any?) -> DefaultMetadata? {
+        guard
+            let analyticsConfigJson = json as? [String: Any?],
+            let json = analyticsConfigJson["defaultMetadata"] as? [String: Any?]
+        else {
+            return nil
+        }
+        let cdnProvider = json["cdnProvider"] as? String
+        let customUserId = json["customUserId"] as? String
+        let customData = analyticsCustomData(json)
+        
+        return DefaultMetadata(
+            cdnProvider: cdnProvider,
+            customUserId: customUserId,
+            customData: customData ?? CustomData()
+        )
     }
 
     /**